requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
Flask-Cors==4.0.0
python-dotenv==1.0.0
pytest==7.4.3
Werkzeug==3.1.6
itsdangerous==2.2.0
click==8.1.6
Jinja2==3.1.2
MarkupSafe==2.1.5
greenlet==3.3.2
pluggy==1.6.0
iniconfig==2.3.0
